backend/requirements.txt
args==0.1.0
asgiref==3.2.10
assertpy==1.0
astroid==2.4.2
attrs==19.3.0
behave==1.2.6
behave-restful==0.3.1
better-exceptions==0.2.2
boto3==1.14.7
botocore==1.17.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
clint==0.5.1
coverage==5.1
cryptography==2.9.2
Django==3.0.7
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-model-utils==4.0.0
django-safedelete==0.5.5
djangorestframework==3.11.0
docutils==0.16
dry-rest-permissions==0.1.10
expects==0.9.0
factory-boy==2.12.0
Faker==4.1.1
glob2==0.7
idna==2.9
importlib-metadata==1.6.1
isort==4.3.21
jmespath==0.10.0
jsonpath==0.82
jsonref==0.2
jsonschema==3.2.0
lazy-object-proxy==1.5.0
Mako==1.1.3
mamba==0.11.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
mysqlclient==1.4.6
olefile==0.46
packaging==20.4
parse==1.15.0
parse-type==0.5.2
pep8==1.7.1
Pillow==7.1.2
pluggy==0.13.1
py==1.8.2
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.3
pytest-bdd==3.4.0
pytest-django==3.9.0
python-dateutil==2.8.1
python-http-client==3.2.7
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
sendgrid==6.3.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.4
wrapt==1.12.1
zipp==3.1.0
